requirements.txt
fastapi
uvicorn[standard]
beanie
motor
pytest-asyncio
httpx
pyjwt[crypto]
pydantic-settings
python-dotenv
